refactor: 어학 로고를 대학 앱 자산으로 정리
