--- apps/web/src/utils/languageUtils.ts
@@ -1,6 +1,0 @@
-export {
-  DEFAULT_LANGUAGE_TEST_LOGO_SRC,
-  formatLanguageTestName,
-  getLanguageTestLogo,
-  languageTestLogoMap as logoMap,
-} from "@solid-connect/ui/language-test-logos";
